v05

+++ reactfront/src/pages/Participantes/ShowDatos.js
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 import Table from 'react-bootstrap/Table';
+import { useParams, useNavigate } from 'react-router-dom';
 import Button from 'react-bootstrap/Button';
 
 const endpoint = 'http://localhost:8000/api';
@@ -8,16 +9,18 @@ const endpoint = 'http://localhost:8000/api';
 const ShowDatos = () => {
     const [datos, setDatos] = useState([]);
     const [error, setError] = useState(null);
+    const { id } = useParams();
+    const navigate = useNavigate(); // Inicializa useNavigate
 
     useEffect(() => {
         getAllDatos();
-    }, []);
+    }, [id]);
 
     const getAllDatos = async () => {
         try {
-            const response = await axios.get(`${endpoint}/datos`);
+            const response = await axios.get(`${endpoint}/datos?with=statusSeleccion`);
             console.log('Datos obtenidos:', response.data);
-            setDatos(response.data);
+            setDatos(response.data.data); // Asegurarse de que `response.data.data` contenga los datos correctamente.
         } catch (error) {
             setError('Error fetching data');
             console.error('Error fetching data:', error);
@@ -25,12 +28,14 @@ const ShowDatos = () => {
     };
 
     const deleteDatos = async (id) => {
-        try {
-            await axios.delete(`${endpoint}/datos/${id}`);
-            getAllDatos();
-        } catch (error) {
-            setError('Error deleting data');
-            console.error('Error deleting data:', error);
+        if (window.confirm('¿Estás seguro de que deseas eliminar este Participante?')) {
+            try {
+                await axios.delete(`${endpoint}/datos/${id}`);
+                getAllDatos();
+            } catch (error) {
+                setError('Error deleting data');
+                console.error('Error deleting data:', error);
+            }
         }
     };
 
@@ -41,43 +46,42 @@ const ShowDatos = () => {
     return (
         <div className="container mt-5">
             <meta name="csrf-token" content="{{ csrf_token() }}"></meta>
-            <h1>Lista de Datos Registrados</h1>
-            <Table striped bordered hover>
+            <h1>Lista de Participantes</h1>
+            <div className="cards-container"></div>
+            <Table striped bordered hover className="rounded-table">
                 <thead>
                     <tr>
                         <th>Cédula</th>
                         <th>Nombres</th>
                         <th>Apellidos</th>
-                        <th>Status</th>
+                        {/* <th>Status</th> */}
                         <th>Acciones</th>
                     </tr>
                 </thead>
                 <tbody>
-                   
-                    {datos?.data?.map((dato) => (
-
+                    {datos.map((dato) => (
                         <tr key={dato.cedula_identidad}>
                             <td>{dato.cedula_identidad}</td>
                             <td>{dato.nombres}</td>
                             <td>{dato.apellidos}</td>
-                            <td>{dato.statusSeleccion ? dato.statusSeleccion.descripcion : 'N/A'}</td>
+                            {/* <td>{dato.status_seleccion?.descripcion || 'N/A'}</td> */}
                             <td>
-                                <Button 
-                                    variant="info" 
-                                    href={`/datos/${dato.cedula_identidad}`}
+                                <Button
+                                    variant="info"
+                                    onClick={() => navigate(`/datos/${dato.cedula_identidad}`)}
                                     className="me-2"
                                 >
                                     Ver más
                                 </Button>
-                                <Button 
-                                    variant="warning" 
-                                    href={`/datos/${dato.cedula_identidad}/edit`}
+                                <Button
+                                    variant="warning"
+                                    onClick={() => navigate(`/datos/${dato.cedula_identidad}/edit`)}
                                     className="me-2"
                                 >
                                     Editar
                                 </Button>
-                                <Button 
-                                    variant="danger" 
+                                <Button
+                                    variant="danger"
                                     onClick={() => deleteDatos(dato.cedula_identidad)}
                                 >
                                     Eliminar
@@ -87,7 +91,11 @@ const ShowDatos = () => {
                     ))}
                 </tbody>
             </Table>
-            <Button variant="success" href="/formulario/create">
+            <Button
+                variant="success"
+                onClick={() => navigate('/formulario/create')}
+                className="mt-3"
+            >
                 Agregar Nuevo Registro
             </Button>
         </div>
